Refactored game handling and updated UI elements
Refactored the game creation and editing logic by moving it to a new `GameService`, simplifying the controller code. Removed the `GameExists` method and `IGenreService` dependency. Updated the method for getting recommended games to `GetReccomendedGamesAsync` and changed the selection logic to randomly pick 8 games. Made game details properties nullable and required. Added `GameService` to the list of services and implemented the `GetLastBoughtGameAsync` method in `LibraryService.cs`. Changed the game key value to a GUID in `OrderService.cs`. Updated UI elements in various `.cshtml` files, including adding a column to the shopping cart table, updating the order completion message, and adding a new partial view for recommended games based on the user's last purchase.